(PA-1995) Use location & version in puppet-runtime.json
Replaces the git ref and url values in puppet-runtime.json with version
and location values instead. This obviates the need to push, tag, and
ship runtime revisions before using them to build an agent; The README
has been updated with instructions on using a local puppet-runtime
output directory as a tarball source instead of Puppet's internal build
sources.