(PA-2158) Add el-7-ppc64 build support for 6.x
This patch adds support for RHEL 7 on big-endian powerpc systems. In particular,
this has been built for a POWER8 environment running RHEL 7.5. Changes are
focused on adding the ppc64 architecture to various locations in the config
files which explicitly check for ppc64le instead.

Note that I do not run any instances of the 6.x series agent in my environment
as the series has yet to be released officially.