(PA-2241) Rewrite the pxp-agent service's AppExit configuration
Previously, NSSM would restart the pxp-agent service if it fails
for any other reason besides an invalid service configuration.
If the failure happens when the service is starting up, then
there's a chance that the restart is stuck in an infinite loop.
We've seen this happen in Puppet 6.

This commit configures NSSM to restart the service on an exit code
of 1, and stop for all other exit codes. An exit code of 1 likely
corresponds to a crash or an unexpected failure, so restarting the
service here makes sense. The other exit codes correspond to failures
in the setup, so stopping the service for those exit codes also makes
sense. Note that pxp-agent returned an exit code of 3 during the
infinite loop scenario, so the changes here should fix that.