(MODULES-8255) switch vendored module git urls from git to https
git:// urls use the default git protocol and are plaintext. To increase
the security during builds, I am switching it to https. This prodives
transport encryption and makes man in the middle attacks harder. The PR
is related to https://tickets.puppetlabs.com/browse/MODULES-8255 but
does not completely resolve it.