(packaging) Move facter/pxp-agent/leatherman back
Facter/pxp-agent/leatherman were all previously pointed at master branch SHAs.
Now that the 6.1.x branch exists we need to move those SHAs back to the
respective heads of the maintenance branches going in to this release.

This means:

* 3.12.x HEAD for facter
* 1.5.x HEAD for leatherman
* 1.10.x HEAD for pxp-agent

All other components are either tagged already, or are moving to new
maintenance branches