(maint) Confine SCL pip test to 64 bit centos
The ensure_SCL_python_pip_works test uses redhat SCL packages to perform a
python 'pip' install so we can test that the pip provider still works with SCL
python.

However, the test uses actual SCL packages on centos and those are only
available for 64 bit platforms.

This commit updates the test to confine to 64 bit centos, and
https://tickets.puppetlabs.com/browse/PA-2386 has been opened to make the test
not rely on actual SCL packages in the future.