(maint) Allow for keeping headers with DEV_BUILD
In the final steps of puppet-agent builds, the 'cleanup' component
removes many files that are unnecessary in finished packages, including
headers for libraries like boost.

This change introduces a new environment variable, $DEV_BUILD, that,
when set, will make vanagon skip the cleanup so that all files
(including headers) are retained in the final package. This is useful
for C++ some development workflows.