(PA-2752) Temporarily confine python3.4 on Jessie
A bug was introduced in the python3.4 Jessie package, which is being
tracked here: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=931044
This causes packaging jobs to fail on Debian 8 (Jessie), since package
devscripts which is installed on this platform depends on python3.4.
Until this gets fixed, we need to confine python3.4 and its dependencies
to the previous version (3.4.2-1).