(PA-2768) Extend manpath instead of overriding it
Whoever implemented MANPATH apparently didn't do it like PATH.

By default, MANPATH is not set and man relies on paths specified in
/etc/man_db.conf. When setting MANPATH to a single path, the default
paths are overrided with this variable.

To make it work we need to prepend a colon to the exported variable, so
that the search path gets extended instead of being replaced.