(maint) Fix repo smoke tests
When looking for the expected packages, the release field was hardcoded
to '1' for all packages. While for most packages this will work, release
packages often have multiple releases over their lifetime, so the test
was failing to pick up the release package. This updates the regex to
include the release field and allow that to be `[0-9]+`.