(maint) Do not copy over libstdc++ and libgcc_s on AIX
I'm not sure what's going on on with the main runtime, but all AIX
builds fail when `cp`-ing the `libgcc_s.a` file to
`/opt/puppetlabs/puppet/lib`, with the following message:

cp: /opt/puppetlabs/puppet/lib/libgcc_s.a: Text file busy

This can be reproduced afterwards by ssh-ing to the target AIX box and
manually running:

cp -p
'/opt/pl-build-tools/lib/gcc/powerpc-ibm-aix6.1.0.0/5.2.0/libgcc_s.a'
'/opt/puppetlabs/puppet/lib/libgcc_s.a'

The failure can be circumvented by adding the `-f` flag to `cp` which
deletes the target file first if it cannot overwrite it.

But why do that when we can drop this entirely, as both `libstdc++.a`
and `libgcc_s.a` are already under `/opt/puppetlabs/puppet/lib`, and
the files are identical.

I don't understand why this happens, as both runtimes are almost
identical except for the ruby version, and master passes while main
fails.... anyway, let's drop this and see what happens.