(PA-3405) Fix puppet resource output matching
PUP-3721[1] included an additional parameter in the `puppet resource`
command output which had the side effect of changing the indentation of
the output. As such, attempting to match `ensure =>` no longer works.

This causes agent upgrades on RPM-based SysV platforms and AIX to no
longer start the puppet service after the upgrade finishes.

Make the regex more permissive in order to account for any number of
spaces between `ensure` and `=>`.

[1] https://github.com/puppetlabs/puppet/commit/7a92874522d6540157168fd025174bba3b816c64