(maint) Replace triple dots in TRAVIS_COMMIT_RANGE with double dots
The TRAVIS_COMMIT_RANGE environment variable uses triple dots to describe the
range[1]. When that is used in `git log base...HEAD`, then it will include all
commits that are not in both A and B. That means we will include commits in the
base branch that are unrelated to the PR.

For now, do what puppet does to convert triple dots to double dots. In the
future, use GITHUB_BASE_REF to determine the range of commits.

[1] https://github.com/travis-ci/travis-ci/issues/4596

(cherry picked from commit f693074fdc3060f9ed6d663aeef1901cc35ff634)