(PA-4646) Add homebrew ruby@x.y dependency from deleted hiera component
The hiera component was deleted in commit 7c931cda2, but the hiera component had
a side effect of installing ruby@x.y when cross compiling macOS. So later
attempts to patch it via `pl-ruby-patch` failed. This moves the build
requirement to the `puppet-runtime` component, similar to how we handle
cross-compiled Linux and Solaris.