(PA-4646) Drop pl-ruby-patch build requirement
The `pl-ruby-patch` component is only necessary when cross compiling, which is
why the `_shared-agent-component` in `puppet-runtime` specifies:

    proj.component 'pl-ruby-patch' if platform.is_cross_compiled?

So do the same in this repo.

Note strictly speaking, the `pl-ruby-patch` component is only needed prior to
calling `gem install --local xxx.gem`. This happens in the `puppet-runtime` repo,
for example, when installing rubygem-ffi. But `puppet-agent` (currently) only
installs puppet, facter and resource_api gems from source using `install.rb`. So
while it's not strictly needed now, it will be in the future if/when we add a
rubygem-xxx component.