(PA-5398) Add openssl-fips 3.0.0 component
Adds fips.so shared library to puppet-agent packages on RHEL 7 & 8 FIPS.

Runs post-install task to install the shared library. This performs a number of
self-tests[1].

The command also generates an MAC for the shared library and records that in
`fipsmodule.cnf` so that the next time the shared library is loaded, openssl
will perform an integrity check, but skip the self test[2].

After the `fipsmodule.cnf` file is created, we install our patched `openssl.cnf`
to load `fipsmodule.cnf` and `fips` provider. However, the `fips` provider does
not contain encoder/decoders need to serialize asymmetric keys like RSA. So we
explicitly activate the `default` provider as recommended in[3]:

    Therefore if the default provider is to be used in conjunction with other
    providers then it must be loaded explicitly.

Finally, we set `default_properties = "fips=yes"` in openssl.cnf as recommended
in[1]:

    The "fips=yes" property can be use to make sure only FIPS approved
    implementations are used for crypto operations. This may also include other
    non-crypto support operations that are not in the FIPS provider, such as
    asymmetric key encoders.

[1] https://www.openssl.org/docs/man3.0/man7/OSSL_PROVIDER-FIPS.html
[2] https://www.openssl.org/docs/man3.0/man5/fips_config.html
[3] https://www.openssl.org/docs/man3.0/man7/OSSL_PROVIDER-default.html