(PA-5399) Enable fips during MSI install/upgrade on FIPS
Add a custom action to run during installation. It runs silently so as to not
display a console window. It executes the same commands are are done for RPM
systems: runs openssl fipsinstall to generate fipsmodule.cnf and copies
openssl-fips.cnf to openssl.cnf. Note the '&' character separates each command
passed to cmd.exe.

Our MSI is configured to always perform a "major upgrade" which removes
everything and installs everything, so the custom action doesn't need to
distinguish between install or upgrade.