(PA-5656) Override gem commands for Solaris 11 SPARC
We natively compile puppet-runtime and pxp-agent-vanagon on Solaris 11 SPARC,
because we were not able to install a new enough gcc on x86_64 to cross compile.

However, we only have 2 Solaris 11 SPARC hosts and we don't need gcc to build
puppet-agent. We just need a runnable ruby to execute `gem install` and `ruby
install.rb`. The latter is used when installing puppet and facter components.

So override the `host_ruby`, etc settings that are inherited from puppet-runtime
to point to pl-build-tools ruby. As a result we're "cross-building"
puppet-agent#main on Solaris 11 SPARC, just like we're doing in 7.x.