(PA-5665) Install OpenCSW GCC on Solaris
Compiling native extensions for the sqlite3 gem fails for Solaris
currently because it attempts to use a compilation flag not understood
by the version of GCC we use on Solaris.

This commit installs the OpenCSW version of GCC on Solaris to enable the
compilation of native extensions for the sqlite3 gem.