(PA-5632) Set SELinux test to high
Prior to this commit, the SELinux acceptance test was set to low
priority, causing the test to be skipped.

However, we've seen multiple regressions on SELinux behavior in
puppet-agent, so we should always run this test to catch them early.

This commit updates the SELinux acceptance test to high priority,
ensuring that it's always run.