Update Ruby version in checks
Prior to this commit, the checks GitHub Action was running on Ruby 2.7.
This version of Ruby went end-of-life in early 2023 and libraries (such
as rubygems-update) have started to drop support, causing errors with
the action.

This commits that Ruby version to 3.2 to match what we're using in
puppet-agent main/8.x.